Issue-1100: Nested & Fallback Claims
As described in Issue-1100 there is current no possibility to mappe values which are nested in other than the root scope. To enable this we change the way how Mapping-Attributes are used. They should understand the "." as object seperator.

Sometimes you like to have a fallback in your mappings. So for displayname for example please take the nickname (if it exists), if it does not exists take the full name ....
To support this, we change the way how Mapping-Attributes are used. They should understand the "|" as an alternative mapping.

The functionality is configurable on the provider, either by occ comand and the ui. As default it is deactivated.

Signed-off-by: Dragonpil <dragonpil@googlemail.com>
Signed-off-by: Julien Veyssier <julien-nc@posteo.net>