Issue-1100: Nested & Fallback Claims
As described in Issue-1100 there is current no possibility to mappe values which are nested in other the the root scope. To enable this we change the way how Mapping-Attributes are used. They should understand the . as object seperator.

Sometimes you like to have a fallback in your mappings. So for displayname for example please take the nickname (if it exists), if it does not exists take the full name ....
To support this, we change the way how Mapping-Attributes are used. They should understand the | as an alternative mapping.

WIP: Testing need to be done still fully, I dont know if I changed everything needed but on my side it works.
Signed-off-by: Dragonpil <dragonpil@googlemail.com>